backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0
feedparser==6.0.11
requests==2.31.0
numpy==1.26.4
neo4j==5.19.0
python-multipart==0.0.9
google-generativeai==0.7.2
duckduckgo-search==8.1.1
newsapi-python==0.2.7
google-search-results==2.4.2
langgraph==0.2.66
